* [filterchange] Cisco Nexus (#594)

Adjust airOS and Nexsus to address event with no HOST but with TZ indicator

* [doc] Palo filter does not support UDP due to limitations of message size (#591)

* [filteradd] Carbon Black Protection CEF format (#590)

* [fix] noise in SC4S logs from goss (#595)

* Troubleshooting/problem >solution (#593) (#596)

* Troubleshooting/problem >solution (#593)

* Troubleshooting guide format problem solution

* Minor chnages

* minor changes

* Revert "Troubleshooting/problem >solution (#593)" (#597)

This reverts commit 20a11167d17ba929bff588c758dd76dd404b0255.


* [filter] improve f5 filter logic (#601)

* [filter] Additional severity levels for f5

* [filter] improve f5 matching

* Fix indents

* Fix indents


* [doc] clarify SC4S_DEST_SPLUNK_HEC_TLS_CA_FILE (#602)

* [doc] clarify SC4S_DEST_SPLUNK_HEC_TLS_CA_FILE

* Clarification edit

* Clarification edit

* Clarify edit 2

* Clarify edit 2

* Clarify edit 3

* Clarify edit 3

* Clarify edit 4

* Clarify edit 4


* [doc] Update troubleshooting guide (#600)

* Update troubleshooting guide

* Unit file updates/home page update (#605)

* Update unit files for proper restart behavior
* Update pathnames for `sleep` and `conntrack`
* Update SC4S docs home page to indicate fully supported status

* Update `entrypoint.sh` HEC check (#607)

* Update `entrypoint.sh` to honor value of `SC4S_DEST_SPLUNK_HEC_TLS_VERIFY` during HEC connectivity check

* [filtermod] Update Citrix to handle malformed AAA (#609)

* [filtermod] Checkpoint doesn't use the correct whitespace (#608)

* [filtermod] CEF time stamp (#612)

False error caused when itterating through multiple time stamps
add support for fractional seconds

* [fix] Incorrect host resolution (#610)

* [fix] Incorrect host resolution

When the log source includes an IP as host resolve using connection IP rather than field IP. When the host resolves to a single name rather than FQDN do not set the host value as this can't be trusted

* Update test_common.py

* Update fix_dns.conf

* [filtermod] Resolve time issue in acs (#613)

* [filtermod] Resolve time issue in acs

ACS more often than not does not send a better time than BSD time field so do not attempt to use it

* Update lp-cisco_acs.conf.tmpl

* [filtermod] cisco date parser issues (#611)

* [filtermod] cisco date parser issues

False error reported when cisco device sends uptime
device reported time errors will now use the indexed field cisco_time_error
Add micro seconds format without year

* Update cisco_syslog.conf

Use cisco time even when wrong

* [filtermod] Fix issue with cp nested syslog (#614)

* [fix] lost brace in merge (#619)

* [filtermod] cisco date parser issues

False error reported when cisco device sends uptime
device reported time errors will now use the indexed field cisco_time_error
Add micro seconds format without year

* Update cisco_syslog.conf

Use cisco time even when wrong

* Update cisco_syslog.conf

* RSA SecurID timestamps (#616)

* Fix time/date parsing in RSA SecurID
* Nit: Should fix all references to this source as `SecurID` (with no "e")

* [fix] revert acs changes (#620)

Revert

* [fix] lookup host by sourceip (#621)

* [fix] lookup host by sourceip

* Update test_common.py

* Update test_cisco_acs.py (#622)

Improve test to avoid flaky ness

* [fix] Fixes #604 (#615)

Do not check client cert in TLS connections as we do not support a reasonable means of authorization checking

* [filtermod] Add program 'iControlPortal.cgi' to f5 list Fixes #568 (#617)

* [doc] prevent conntrack from halting start (#618)

* CEF:  Imperva WAF timestamp parsing (#624)

* CEF:  Imperva WAF timestamp parsing fix

* Fixed Cisco WSA python test (#623)

* Fixed Cisco WSA python test

* Update local container guidance (#625)

* Add item to remove pull from unit file in local container guidance
* Clarify use of internal (syslog-ng based) load balancing vs. external LB

* [feature] Alpha Documentation and changes to support microk8s

This is alpha level support and rough docs for k8s as a runtime

* [filtermod] Resolve f5 UTF issue with glob (#626)

+++ package/sbin/entrypoint.sh
@@ -48,35 +48,49 @@ trap 'kill ${!}; term_handler' SIGTERM
 
 mkdir -p /opt/syslog-ng/etc/conf.d/local/context/
 mkdir -p /opt/syslog-ng/etc/conf.d/local/config/
+
+
 cp /opt/syslog-ng/etc/context_templates/* /opt/syslog-ng/etc/conf.d/local/context
 for file in /opt/syslog-ng/etc/conf.d/local/context/*.example ; do cp --verbose -n $file ${file%.example}; done
+if [ "$SC4S_RUNTIME_ENV" == "k8s" ]
+then
+  mkdir -p /opt/syslog-ng/etc/conf.d/configmap/context/
+  mkdir -p /opt/syslog-ng/etc/conf.d/configmap/config/
+  # Add new entries
+  temp_file=$(mktemp)
+  awk '{print $0}' /opt/syslog-ng/etc/conf.d/configmap/context/splunk_metadata.csv /opt/syslog-ng/etc/context_templates/splunk_metadata.csv.example | grep -v '^#' | sort -b -t ',' -k1,2 -u  > $temp_file
+  cp -f $temp_file /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv
 
-# splunk_index.csv updates
-# Remove comment headers from existing config
-touch /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv
-if [ -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv ]; then
-    LEGACY_SPLUNK_INDEX_FILE=/opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv
-fi
-# Add new entries
-temp_file=$(mktemp)
-awk '{print $0}' ${LEGACY_SPLUNK_INDEX_FILE} /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv /opt/syslog-ng/etc/context_templates/splunk_metadata.csv.example | grep -v '^#' | sort -b -t ',' -k1,2 -u  > $temp_file
-cp -f $temp_file /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv
-# We don't need this file any longer
-rm -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv.example || true 
-if [ -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv ]; then
-    cp -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv /opt/syslog-ng/etc/conf.d/local/context/splunk_index.deprecated
-    rm /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv
-fi
-cp --verbose -R -f /opt/syslog-ng/etc/local_config/* /opt/syslog-ng/etc/conf.d/local/config/
+else
+  # splunk_index.csv updates
+  # Remove comment headers from existing config
+  touch /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv
+  if [ -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv ]; then
+      LEGACY_SPLUNK_INDEX_FILE=/opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv
+  fi
+
+  # Add new entries
+  temp_file=$(mktemp)
+  awk '{print $0}' ${LEGACY_SPLUNK_INDEX_FILE} /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv /opt/syslog-ng/etc/context_templates/splunk_metadata.csv.example | grep -v '^#' | sort -b -t ',' -k1,2 -u  > $temp_file
+  cp -f $temp_file /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv
+  # We don't need this file any longer
+  rm -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv.example || true 
+  if [ -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv ]; then
+      cp -f /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv /opt/syslog-ng/etc/conf.d/local/context/splunk_index.deprecated
+      rm /opt/syslog-ng/etc/conf.d/local/context/splunk_index.csv
+  fi
+  cp --verbose -R -f /opt/syslog-ng/etc/local_config/* /opt/syslog-ng/etc/conf.d/local/config/
+fi  
 mkdir -p /opt/syslog-ng/var/log
 
 # Test HEC Connectivity
 if [ "$SC4S_DEST_SPLUNK_HEC_GLOBAL" != "no" ]
 then
   HEC=$(echo '{{- getenv "SPLUNK_HEC_URL" | strings.ReplaceAll "/services/collector" "" | strings.ReplaceAll "/event" "" | regexp.ReplaceLiteral "[, ]+" "/services/collector/event " }}/services/collector/event' | gomplate | cut -d' ' -f 1)
+  NO_VERIFY=$(echo '{{- if not (conv.ToBool (getenv "SC4S_DEST_SPLUNK_HEC_TLS_VERIFY" "yes")) }}-k{{- end}}' | gomplate)
   SC4S_DEST_SPLUNK_HEC_FALLBACK_INDEX=$(cat /opt/syslog-ng/etc/conf.d/local/context/splunk_metadata.csv | grep ',index,' | grep sc4s_events | cut -d, -f 3)
   export SC4S_DEST_SPLUNK_HEC_FALLBACK_INDEX
-  if curl -s -S -k "${HEC}?/index=${SC4S_DEST_SPLUNK_HEC_FALLBACK_INDEX}" -H "Authorization: Splunk ${SPLUNK_HEC_TOKEN}" -d '{"event": "HEC TEST EVENT", "sourcetype": "SC4S:PROBE"}' 2>&1 | grep -v '{"text":"Success","code":0}'
+  if curl -s -S ${NO_VERIFY} "${HEC}?/index=${SC4S_DEST_SPLUNK_HEC_FALLBACK_INDEX}" -H "Authorization: Splunk ${SPLUNK_HEC_TOKEN}" -d '{"event": "HEC TEST EVENT", "sourcetype": "SC4S:PROBE"}' 2>&1 | grep -v '{"text":"Success","code":0}'
   then
     echo -e "SC4S_ENV_CHECK_HEC: Invalid Splunk HEC URL, invalid token, or other HEC connectivity issue.\nStartup will continue to prevent data loss if this is a transient failure."
   else
